Refactor PatchServiceProperties package location
Resolves:
PSC-122
